feat(error): moving constant error messages into separate file
